# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  dto.ResetPasswordRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      new_password:
        type: string
      token:
        type: string
    required:
    - confirm_password
    - email
    - new_password
    - token
    type: object
  dto.UserMeResponse:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      full_name:
        type: string
      phone_number:
        type: string
      postal_code:
        type: string
      province:
        type: string
    type: object
  responseswagger.BadRequestResponse:
    properties:
      message:
        example: format json tidak valid
        type: string
    type: object
  responseswagger.ConflictResponse:
    properties:
      message:
        example: data sudah terdaftar
        type: string
    type: object
  responseswagger.ForgotPasswordSuccessResponse:
    properties:
      message:
        example: tautan pemulihan kata sandi telah dikirim ke email
        type: string
    type: object
  responseswagger.InternalServerErrorResponse:
    properties:
      message:
        example: internal server error
        type: string
    type: object
  responseswagger.LoginSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginResponse'
      message:
        example: login berhasil
        type: string
    type: object
  responseswagger.LogoutSuccessResponse:
    properties:
      message:
        example: logout berhasil, sesi telah dihapus
        type: string
    type: object
  responseswagger.RegisterSuccessResponse:
    properties:
      message:
        example: pendaftaran berhasil
        type: string
    type: object
  responseswagger.ResetPasswordSuccessResponse:
    properties:
      message:
        example: kata sandi berhasil diperbarui, silakan login kembali
        type: string
    type: object
  responseswagger.UserMeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserMeResponse'
      message:
        example: data profil berhasil diambil
        type: string
    type: object
  responseswagger.ValidationFailedResponse:
    properties:
      errors:
        properties:
          email:
            example:
            - format email salah
            items:
              type: string
            type: array
          full_name:
            example:
            - nama tidak valid
            items:
              type: string
            type: array
          password:
            example:
            - minimal 6 karakter
            items:
              type: string
            type: array
        type: object
      message:
        example: validasi gagal
        type: string
    type: object
  responseswagger.VerifyEmailSuccessResponse:
    properties:
      message:
        example: verifikasi akun berhasil
        type: string
    type: object
info:
  contact:
    email: ekokurniawaann@gmail.com
    name: Eko Kurniawan
  description: Dokumentasi API untuk Aplikasi FORTISFIT
  termsOfService: http://swagger.io/terms/
  title: Backend Skripsi API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      description: Mengirimkan tautan kriptografis berdurasi 15 menit ke kotak masuk
        email jika email terdaftar.
      parameters:
      - description: Payload Email Pemulihan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseswagger.ForgotPasswordSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responseswagger.ValidationFailedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      summary: Minta Link Lupa Password
      tags:
      - Auth
  /auth/login:
    post:
      description: Autentikasi email dan password pengguna untuk mendapatkan token
        akses JWT.
      parameters:
      - description: Kredensial Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseswagger.LoginSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responseswagger.ValidationFailedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      summary: Login Pengguna
      tags:
      - Auth
  /auth/logout:
    post:
      description: Memasukkan token JWT yang sedang digunakan ke dalam daftar hitam
        Redis agar tidak bisa digunakan kembali sebelum masa kedaluwarsanya habis.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseswagger.LogoutSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout Pengguna
      tags:
      - Auth
  /auth/me:
    get:
      description: Mengambil data profil pengguna yang sedang login berdasarkan token
        JWT yang dikirim di Header Authorization.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseswagger.UserMeSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil Profil Pengguna (Me)
      tags:
      - Auth
  /auth/register:
    post:
      description: Registrasi akun pengguna baru ke sistem FortisFit.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responseswagger.RegisterSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/responseswagger.ConflictResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responseswagger.ValidationFailedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      summary: Register Akun Baru
      tags:
      - Auth
  /auth/reset-password:
    post:
      description: Mengubah kata sandi lama menjadi kata sandi baru jika token email
        valid dan belum kedaluwarsa.
      parameters:
      - description: Payload Password Baru dan Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseswagger.ResetPasswordSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responseswagger.ValidationFailedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      summary: Eksekusi Reset Password Baru
      tags:
      - Auth
  /auth/verify:
    get:
      description: Memvalidasi token dari email untuk mengubah status is_verified
        menjadi true.
      parameters:
      - description: Email Pengguna
        in: query
        name: email
        required: true
        type: string
      - description: Token Verifikasi Kriptografis
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responseswagger.VerifyEmailSuccessResponse'
        "400":
          description: Token atau email tidak valid
          schema:
            $ref: '#/definitions/responseswagger.BadRequestResponse'
        "422":
          description: Gagal lolos validasi format
          schema:
            $ref: '#/definitions/responseswagger.ValidationFailedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responseswagger.InternalServerErrorResponse'
      summary: Verify User Email
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"